legger til feature-toggle for ny visning av journalføring. Validering av journalføring skal gjøres i egen fil
